--- a/db/migrate/20170905044538_add_index_id_account_id_activity_type_on_notifications.rb
+++ b/db/migrate/20170905044538_add_index_id_account_id_activity_type_on_notifications.rb
@@ -1,0 +1,5 @@
+class AddIndexIdAccountIdActivityTypeOnNotifications < ActiveRecord::Migration[5.1]
+  def change
+    add_index :notifications, [:id, :account_id, :activity_type], order: { id: :desc }
+  end
+end