In friends.nico, we had these slow queries.

```
mastodon-production=>  SELECT pid, usename, query, query_start FROM pg_stat_activity WHERE state='active' ORDER BY query_start ASC;
  pid  | usename  |                                                                                                                                                          query                                                                                                                                                          |          query_start          
-------+----------+-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-------------------------------
 10405 | mastodon | SELECT  "notifications"."id", "notifications"."updated_at", "notifications"."activity_type", "notifications"."activity_id" FROM "notifications" WHERE "notifications"."account_id" = $1 AND "notifications"."activity_type" IN ('Mention', 'Status', 'Follow', 'Favourite') ORDER BY "notifications"."id" DESC LIMIT $2 | 2017-08-28 09:21:12.621383+00
 10355 | mastodon | SELECT  "notifications"."id", "notifications"."updated_at", "notifications"."activity_type", "notifications"."activity_id" FROM "notifications" WHERE "notifications"."account_id" = $1 AND "notifications"."activity_type" IN ('Mention', 'Status', 'Follow', 'Favourite') ORDER BY "notifications"."id" DESC LIMIT $2 | 2017-08-28 09:21:13.085558+00
 10766 | mastodon | SELECT  "notifications"."id", "notifications"."updated_at", "notifications"."activity_type", "notifications"."activity_id" FROM "notifications" WHERE "notifications"."account_id" = $1 AND "notifications"."activity_type" IN ('Mention', 'Status', 'Follow', 'Favourite') ORDER BY "notifications"."id" DESC LIMIT $2 | 2017-08-28 09:21:13.106323+00
 10358 | mastodon | SELECT  "notifications"."id", "notifications"."updated_at", "notifications"."activity_type", "notifications"."activity_id" FROM "notifications" WHERE "notifications"."account_id" = $1 AND "notifications"."activity_type" IN ('Mention', 'Status', 'Follow', 'Favourite') ORDER BY "notifications"."id" DESC LIMIT $2 | 2017-08-28 09:21:13.116587+00
 10356 | mastodon | SELECT  "notifications"."id", "notifications"."updated_at", "notifications"."activity_type", "notifications"."activity_id" FROM "notifications" WHERE "notifications"."account_id" = $1 AND "notifications"."activity_type" IN ('Mention', 'Status', 'Follow', 'Favourite') ORDER BY "notifications"."id" DESC LIMIT $2 | 2017-08-28 09:21:13.129636+00
 10378 | mastodon | SELECT  "notifications"."id", "notifications"."updated_at", "notifications"."activity_type", "notifications"."activity_id" FROM "notifications" WHERE "notifications"."account_id" = $1 AND "notifications"."activity_type" IN ('Mention', 'Status', 'Follow', 'Favourite') ORDER BY "notifications"."id" DESC LIMIT $2 | 2017-08-28 09:21:14.307262+00
 10366 | mastodon | SELECT  "notifications"."id", "notifications"."updated_at", "notifications"."activity_type", "notifications"."activity_id" FROM "notifications" WHERE "notifications"."account_id" = $1 AND "notifications"."activity_type" IN ('Mention', 'Status', 'Follow', 'Favourite') ORDER BY "notifications"."id" DESC LIMIT $2 | 2017-08-28 09:21:15.023977+00
 11429 | mastodon | SELECT  "notifications"."id", "notifications"."updated_at", "notifications"."activity_type", "notifications"."activity_id" FROM "notifications" WHERE "notifications"."account_id" = $1 AND "notifications"."activity_type" IN ('Mention', 'Status', 'Follow', 'Favourite') ORDER BY "notifications"."id" DESC LIMIT $2 | 2017-08-28 09:21:15.781861+00
 10441 | mastodon | SELECT  "notifications"."id", "notifications"."updated_at", "notifications"."activity_type", "notifications"."activity_id" FROM "notifications" WHERE "notifications"."account_id" = $1 AND "notifications"."activity_type" IN ('Mention', 'Status', 'Follow', 'Favourite') ORDER BY "notifications"."id" DESC LIMIT $2 | 2017-08-28 09:21:16.420561+00
 10359 | mastodon | SELECT  "notifications"."id", "notifications"."updated_at", "notifications"."activity_type", "notifications"."activity_id" FROM "notifications" WHERE "notifications"."account_id" = $1 AND "notifications"."activity_type" IN ('Mention', 'Status', 'Follow', 'Favourite') ORDER BY "notifications"."id" DESC LIMIT $2 | 2017-08-28 09:21:17.830487+00
 17131 | mastodon | SELECT pid, usename, query, query_start FROM pg_stat_activity WHERE state='active' ORDER BY query_start ASC;                                                                                                                                                                                                            | 2017-08-28 09:21:18.373457+00
(11 rows)

mastodon-production=>
```

These queries took about 1000 ～ 8000ms.
We got bellow explain analyze.

before)
```
mastodon-production=> explain analyze SELECT  "notifications"."id", "notifications"."updated_at", "notifications"."activity_type", "notifications"."activity_id" FROM "notifications" WHERE "notifications"."account_id" = ***** AND "notifications"."activity_type" IN ('Mention', 'Status', 'Follow', 'Favourite') ORDER BY "notifications"."id" DESC LIMIT 15;
                                                                           QUERY PLAN                                                                           
----------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.43..5155.13 rows=15 width=29) (actual time=910.979..3401.273 rows=15 loops=1)
   ->  Index Scan Backward using notifications_pkey on notifications  (cost=0.43..486603.34 rows=1416 width=29) (actual time=910.974..3401.245 rows=15 loops=1)
         Filter: ((account_id = *****) AND ((activity_type)::text = ANY ('{Mention,Status,Follow,Favourite}'::text[])))
         Rows Removed by Filter: 7990584
 Planning time: 0.140 ms
 Execution time: 3401.305 ms
(6 rows)

mastodon-production=>
```

Then, I added index(account_activity_test5) on notification table.

```
mastodon-production=> \d notifications
                                        Table "public.notifications"
     Column      |            Type             |                         Modifiers                          
-----------------+-----------------------------+------------------------------------------------------------
 id              | integer                     | not null default nextval('notifications_id_seq'::regclass)
 account_id      | integer                     | 
 activity_id     | bigint                      | 
 activity_type   | character varying           | 
 created_at      | timestamp without time zone | not null
 updated_at      | timestamp without time zone | not null
 from_account_id | integer                     | 
Indexes:
    "notifications_pkey" PRIMARY KEY, btree (id)
    "account_activity" UNIQUE, btree (account_id, activity_id, activity_type)
    "account_activity_test5" btree (id DESC, account_id, activity_type)
    "index_notifications_on_activity_id_and_activity_type" btree (activity_id, activity_type)
Foreign-key constraints:
    "fk_rails_1c0a19e3ee" FOREIGN KEY (account_id) REFERENCES accounts(id) ON DELETE CASCADE
    "fk_rails_a9df2835a1" FOREIGN KEY (from_account_id) REFERENCES accounts(id) ON DELETE CASCADE

mastodon-production=> 
```

After added index, the query has lower CPU Usage and faster.

after)
```
mastodon-production=>  explain analyze SELECT  "notifications"."id", "notifications"."updated_at", "notifications"."activity_type", "notifications"."activity_id" FROM "notifications" WHERE "notifications"."account_id" = ***** AND "notifications"."activity_type" IN ('Mention', 'Status', 'Follow', 'Favourite') ORDER BY "notifications"."id" DESC LIMIT 15;
                                                                      QUERY PLAN                                                                       
-------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.56..71.57 rows=15 width=29) (actual time=0.019..0.056 rows=15 loops=1)
   ->  Index Scan using account_activity_test5 on notifications  (cost=0.56..357441.37 rows=75501 width=29) (actual time=0.018..0.053 rows=15 loops=1)
         Index Cond: (account_id = *****)
         Filter: ((activity_type)::text = ANY ('{Mention,Status,Follow,Favourite}'::text[]))
 Planning time: 0.171 ms
 Execution time: 0.080 ms
(6 rows)

mastodon-production=>
```

CPU Usage)
![image](https://user-images.githubusercontent.com/24884114/29913497-57c8a49e-8e70-11e7-936c-91741840ddf2.png)