@beatrix-bitrot 2017-08-31
curious how many notifications in the DB and the size of this index!

also wonder if we have any blocking operations involving notification creation and if their speed is affected 🤔 

anyway i like it 👌