@takayamaki 2017-09-01
Our notifications table is 8757870 rows / 747MB, and index(account_activity_test5) is 338MB.
I feel that the index is large, but database planner selected it from our tried some indexes.

Index creation was done during maintenance for updating to v1.5.1.
At that time it took about 15 secs.

I have never seen record creation as live queries, so I don't know how much its speed will be affected.