@nightpool 2017-09-01
If this index takes 15 seconds to complete, it really should be made concurrently so it doesn't require downtime, right?

@beatrix-bitrot my understanding is that all *local* notifications are done in a blocking fashion, or at least used to be?