@takayamaki 2017-09-01
I tried index creation in my another instance(imastodon.net), it took about 2 secs.
imastodon.net's notification table is 422597 rows / 132 MB.

I feel that the huge notification table is a unique of friends.nico.
I think that it is unnecessary for most instances, but it better to attach a concurrently option?

```
mastodon=> \d notifications
                                        Table "public.notifications"
     Column      |            Type             |                         Modifiers                          
-----------------+-----------------------------+------------------------------------------------------------
 id              | integer                     | not null default nextval('notifications_id_seq'::regclass)
 account_id      | integer                     | 
 activity_id     | bigint                      | 
 activity_type   | character varying           | 
 created_at      | timestamp without time zone | not null
 updated_at      | timestamp without time zone | not null
 from_account_id | integer                     | 
Indexes:
    "notifications_pkey" PRIMARY KEY, btree (id)
    "account_activity" UNIQUE, btree (account_id, activity_id, activity_type)
    "index_notifications_on_activity_id_and_activity_type" btree (activity_id, activity_type)
Foreign-key constraints:
    "fk_rails_1c0a19e3ee" FOREIGN KEY (account_id) REFERENCES accounts(id) ON DELETE CASCADE
    "fk_rails_a9df2835a1" FOREIGN KEY (from_account_id) REFERENCES accounts(id) ON DELETE CASCADE

mastodon=> explain analyze SELECT  "notifications"."id", "notifications"."updated_at", "notifications"."activity_type", "notifications"."activity_id" FROM "notifications" WHERE "notifications"."account_id" = 1 AND "notifications"."activ
ity_type" IN ('Mention', 'Status', 'Follow', 'Favourite') ORDER BY "notifications"."id" DESC LIMIT 15;
                                                                        QUERY PLAN                                                                        
----------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.42..148.24 rows=15 width=29) (actual time=1.412..2.078 rows=15 loops=1)
   ->  Index Scan Backward using notifications_pkey on notifications  (cost=0.42..60378.16 rows=6127 width=29) (actual time=1.410..2.070 rows=15 loops=1)
         Filter: ((account_id = 1) AND ((activity_type)::text = ANY ('{Mention,Status,Follow,Favourite}'::text[])))
         Rows Removed by Filter: 1649
 Planning time: 0.300 ms
 Execution time: 2.097 ms
(6 rows)

mastodon=> SELECT     relname,     reltuples as rows,     (relpages * 8192) as bytes FROM     pg_class WHERE     relname = 'notifications';
    relname    |  rows  |   bytes   
---------------+--------+-----------
 notifications | 422597 | 137969664
(1 row)

mastodon=> CREATE INDEX "account_activity_test" ON "notifications" ("id" DESC, "account_id", "activity_type");
CREATE INDEX
mastodon=> explain analyze SELECT  "notifications"."id", "notifications"."updated_at", "notifications"."activity_type", "notifications"."activity_id" FROM "notifications" WHERE "notifications"."account_id" = 1 AND "notifications"."activity_type" IN ('Mention', 'Status', 'Follow', 'Favourite') ORDER BY "notifications"."id" DESC LIMIT 15;
                                                                     QUERY PLAN                                                                      
-----------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.42..65.92 rows=15 width=29) (actual time=0.020..0.101 rows=15 loops=1)
   ->  Index Scan using account_activity_test on notifications  (cost=0.42..27223.72 rows=6235 width=29) (actual time=0.019..0.092 rows=15 loops=1)
         Index Cond: (account_id = 1)
         Filter: ((activity_type)::text = ANY ('{Mention,Status,Follow,Favourite}'::text[]))
 Planning time: 0.285 ms
 Execution time: 0.121 ms
(6 rows)

mastodon=>
```