@takayamaki 2017-09-05
I fixed conflict.